chore: initial refactor to use ramda
